requirements.txt
google-adk
fastapi
uvicorn[standard]
python-dotenv
litellm
boto3
pydantic
sqlalchemy
psycopg2-binary